Changes to Main.py, add Class.py and update README.md
