+++ README.md
@@ -7,3 +7,6 @@ This is the Read Me file added to the Git Repository for the Week 2 ePortfilio t
 # Test Branch
 
 Cut a branch, and then updated the README.md  Further updates to be addded.
+
+#Update
+Split Class out into seperate file.
